src/gateway/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
passlib==1.7.4
bcrypt==4.0.1
requests
slowapi
prometheus-client
structlog
python-json-logger
pydantic[email]

